requirements.txt
celery==5.3.6
Django==4.2.8
django-celery-beat==2.5.0
django-celery-results==2.5.1
gunicorn==21.2.0
psycopg[binary]==3.1.17 
python-decouple==3.8
redis==5.0.1
whitenoise==6.6.0
py3-validate-email==1.0.5
